test: emoji service w/o posting passed naive test, some modifications followed

+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/service/EmojiService.kt
@@ -1,5 +1,6 @@
 package com.goliath.emojihub.springboot.service
 
+import com.goliath.emojihub.springboot.common.CustomHttp404
 import com.goliath.emojihub.springboot.dao.EmojiDao
 import com.goliath.emojihub.springboot.dto.emoji.EmojiDto
 import com.goliath.emojihub.springboot.dto.emoji.PostEmojiRequest
@@ -12,6 +13,8 @@ class EmojiService (private val emojiDao: EmojiDao) {
     }
 
     fun getEmoji(emojiId: String): EmojiDto? {
+        if (emojiDao.existsEmoji(emojiId).not())
+            throw CustomHttp404("Emoji doesn't exist.")
         return emojiDao.getEmoji(emojiId)
     }
 
@@ -20,14 +23,18 @@ class EmojiService (private val emojiDao: EmojiDao) {
     }
 
     fun saveEmoji(userId: String, emojiId: String) {
+        // TODO: check if userId, emojiId exists
         emojiDao.saveEmoji(userId, emojiId)
     }
 
     fun unSaveEmoji(userId: String, emojiId: String) {
+        // TODO: check if userId, emojiId exists
         emojiDao.unSaveEmoji(userId, emojiId)
     }
 
     fun deleteEmoji(emojiId: String) {
+        if (emojiDao.existsEmoji(emojiId).not())
+            throw CustomHttp404("Emoji doesn't exist.")
         emojiDao.deleteEmoji(emojiId)
     }
 }
